media drag overlay cleanup
